Crate quackngo

A Duck Duck Go InstantAnswer API library written in rust.

Copyright 2016 Mark Sta Ana.

Licensed under the Apache License, Version 2.0 at your option. This file may not be copied, modified, or distributed except according to those terms.

Usage

This crate is on crates.io and can be used by adding quack to the dependencies in your project's Cargo.toml.

[dependencies]
quackngo = "0.1"

and this to your crate root:

extern crate quackngo;

use quackngo::Quack;
println!("{:#?}", Quack::new("!imdb+rushmore"));

Structs

Icon
child struct of InstantAnswer

InstantAnswer
DuckDuckGo InstantAnswer

Quack
Main struct

RelatedTopic
child struct of InstantAnswer

Topic
child struct of InstantAnswer