Struct quackngo::InstantAnswer

pub struct InstantAnswer {
    pub abstract_rich: String,
    pub abstract_text: String,
    pub abstract_source: String,
    pub abstract_url: String,
    pub image: String,
    pub heading: String,
    pub answer: String,
    pub redirect: String,
    pub answer_type: String,
    pub definition: String,
    pub definition_source: String,
    pub definition_url: String,
    pub related_topics: Vec<RelatedTopic>,
    pub results: Vec<RelatedTopic>,
    pub response_type: String,
}

DuckDuckGo InstantAnswer

Fields
abstract_rich: String abstract_text: String

topic summary (with no HTML)

abstract_source: String

name of Abstract source

abstract_url: String

deep link to expanded topic page in AbstractSource

image: String

link to image that goes with Abstract

heading: String

name of topic that goes with Abstract

answer: String

instant answer

redirect: String

!bang redirect URL

answer_type: String

type of Answer, e.g. calc, color, digest, info, ip, iploc, phone, pw, rand, regexp, unicode, upc, or zip

definition: String

dictionary definition (may differ from Abstract)

definition_source: String

name of Definition source

definition_url: String

deep link to expanded definition page in DefinitionSource

related_topics: Vec<RelatedTopic>

Vec<RelatedTopics> of internal links to related topics associated with Abstract

results: Vec<RelatedTopic> response_type: String

response category, i.e. A (article), D (disambiguation), C (category), N (name), E (exclusive), or nothing.